% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCIP_filter.R
\name{PCIP_filter}
\alias{PCIP_filter}
\title{Filter Pairwise mApping File produced by Minimap2}
\usage{
PCIP_filter(minimap2PAF = NULL, targetName = NULL)
}
\arguments{
\item{minimap2PAF}{tibble. Pairwise mApping file generated by minimap2 and loaded with \code{readPairwiseAlignmentFile}.}

\item{targetName}{character. Name of the TARGET chromosome.}
}
\value{
PCIP filtered and formated PAF tibble. Field description:
\itemize{
\item seqnames (character). Chromosome name.
\item start (integer). Substring mapping start position.
\item end (integer). Substring mapping end position.
\item readStart_position (integer). Start position of the substring within the read sequence.
\item readEnd_position (integer). End position of the substring within the read sequence.
\item width (integer). Width of the substring.
\item strand (character). Mapping strand (+ or -).
\item ligation (character). TARGET-HOST structure of the read as binary code (0 = HOST and 1 = TARGET).
\item mapGap (integer). Length of the gap between the end of this mapped substring and the next mapped read substring.
\item readID (character). ID of the read.
\item qLength (integer). Total length of the read.
\item target (logical). TARGET chromosome.
\item meanGap (integer). Alignment gaps are frequent between substrings in nanopore. Reports the mean gap between substrings of a read.
\item sdGap (integer). Standard deviation of the inter-substring gaps.
}
}
\description{
Filter out secundary alignments and select TARGET-HOST chimeric reads.
In PAF format each line represents the mapping information of a read substring.
Explore the read structure. PCIP-seq reads should contains chimeric viral (1) - host (0) sequences. Only reads matching the following patterns are selected: 101, 01, 10, 1001.
}
\author{
Vincent Hahaut
}
\keyword{PCIP}
