% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCIP_statistics.R
\name{PCIP_statistics}
\alias{PCIP_statistics}
\title{get Statistics for PCIP-seq}
\usage{
PCIP_statistics(pathToResults = NULL, ID = NULL)
}
\arguments{
\item{pathToResults}{character. Path/to/results/}

\item{ID}{character. ID of the sample}
}
\value{
Write a statistics report in Path/to/results/$ID-statistics.txt
}
\description{
Extract some useful statistics about the run
}
\author{
Vincent Hahaut
}
\keyword{PCIP}
