% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPairwiseAlignmentFile.R
\name{readPairwiseAlignmentFile}
\alias{readPairwiseAlignmentFile}
\title{Read Pairwise Alignment File (PAF) From Minimap2}
\usage{
readPairwiseAlignmentFile(alignFile = NULL)
}
\arguments{
\item{alignFile}{Path to minimap2 PAF file}
}
\description{
Read the first 18th columns from pairwise alignment (PAF) produced by minimap2.

For more information about the PAF format refers to: \href{https://github.com/lh3/miniasm/blob/master/PAF.md}{Minimap2 PAF format}
}
\author{
Vincent Hahaut
}
\keyword{PCIP}
