% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFASTQ.R
\name{splitFASTQ}
\alias{splitFASTQ}
\title{splitFASTQ}
\usage{
splitFASTQ(breakpoints.path = NULL, readID.cleaned.path = NULL,
  mode = "breakpoints", PAF.path = NULL, maxgap = 50,
  out.prefix = NULL)
}
\arguments{
\item{mode}{character. "FULL" (just split the reads by associated IS) or "MAPPED" (split the reads and only keep the mapped parts)}

\item{fastq.path}{character. SAM File path}
}
\value{
a tibble containing the .align results
}
\description{

}
\examples{
readSAMFile(samFile="my/path/to/samFile.sam")

}
\author{
Vincent Hahaut
}
\keyword{PCIP}
